# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ticketing application with restful API powered by gin",
        "title": "CHUBA TICKITZ",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/users/": {
            "get": {
                "security": [
                    {
                        "JWTtoken": []
                    }
                ],
                "description": "Get details user, gt data by known id user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "get profile user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserDetailResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWTtoken": []
                    }
                ],
                "description": "Update user and show new updated data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update registerd user",
                "parameters": [
                    {
                        "description": "Data new user",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/avatar": {
            "patch": {
                "security": [
                    {
                        "JWTtoken": []
                    }
                ],
                "description": "Update user and show new updated data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Update avatar registerd user",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Upload good image",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BadRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Example bad request error..."
                },
                "is_success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Example Internal server error..."
                },
                "is_success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.ProfileResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/models.User"
                },
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Example message success..."
                },
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "point": {
                    "type": "integer"
                },
                "profile_path": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.UserDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/models.User"
                },
                "is_success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Example message success..."
                },
                "page": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "JWTtoken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}